@@ -1,36 +1,53 @@
 {
   "name": "react-hardware",
-  "version": "0.9.0",
-  "description": "React bindings for your hardware through firmata.",
-  "main": "index.js",
+  "version": "0.4.0",
+  "description": "React firmata bindings.",
+  "main": "todo",
   "scripts": {
-    "test": "BABEL_JEST_STAGE=0 jest"
+    "lint": "eslint .",
+    "test": "./scripts/test",
+    "posttest": "npm run lint",
+    "build": "echo \"Error: no build command yet\""
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/iamdustan/react-hardware"
   },
+  "files": [
+    "package.json",
+    "README.md"
+  ],
   "keywords": [
+    "react",
     "reactjs",
     "hardware",
-    "arduino"
+    "arduino",
+    "firmata"
   ],
-  "author": "Dustan Kasten <dustan@skookum.com>",
+  "author": "Dustan Kasten <dustan.kasten@gmail.com>",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/iamdustan/react-hardware/issues"
   },
   "homepage": "https://github.com/iamdustan/react-hardware",
   "dependencies": {
-    "react": "^0.13.2"
+    "fbjs": "0.4.0",
+    "firmata": "^0.8.0",
+    "serialport": "^2.0.5",
+    "setimmediate": "^1.0.4"
   },
   "devDependencies": {
-    "babel": "^5.2.15",
-    "babel-jest": "^5.2.0"
-  },
-  "jest": {
-    "scriptPreprocessor": "<rootDir>/node_modules/babel-jest",
-    "testFileExtensions": ["js"],
-    "moduleFileExtensions": ["js", "json"]
+    "babel-cli": "^6.2.0",
+    "babel-register": "^6.2.0",
+    "chokidar": "^1.3.0",
+    "commander": "^2.9.0",
+    "eslint": "^1.10.3",
+    "eslint-plugin-react": "^3.11.3",
+    "fbjs-scripts": "^0.5.0",
+    "jasmine": "^2.3.2",
+    "jasmine-es6": "0.0.18",
+    "jasmine-spec-reporter": "^2.4.0",
+    "mock-firmata": "0.0.4",
+    "react": "^0.14.3"
   }
 }
